.history/requirements_20240119123143.txt
absl-py
asgiref
astroid
astunparse
cachetools
certifi
charset-normalizer
click
cloudpickle
# contourpy==1.0.7
cycler
dask
# dataclasses-json
dill
distributed
Django
flatbuffers
fonttools
fsspec
gast
google-auth
google-auth-oauthlib
google-pasta
grpcio
h5py
idna
importlib-metadata
install
intervaltree
isort
Jinja2
joblib
kaleido
keras
kiwisolver
lazy-object-proxy
libclang
# libcst
lightgbm
llvmlite
locket
Markdown
MarkupSafe
marshmallow
marshmallow-enum
matplotlib
mccabe
msgpack
mypy
mypy-extensions
numpy
oauthlib
opt-einsum
packaging
pandas
partd
patsy
Pillow
platformdirs
plotly
# protobuf
psutil
psycopg2-binary
pyasn1
pyasn1-modules
Pygments
pylint
pyparsing
# PyQt3D
# PyQt5
PyQt5-sip
# PyQtChart
# PyQtDataVisualization
# PyQtNetworkAuth
# PyQtPurchasing
# PyQtWebEngine
pyre-check
pyre-extensions
python-dateutil
pytz
PyYAML
requests
requests-oauthlib
rsa
scikit-learn
scipy
seaborn
shap
sip
six
# sklearn
# slicer
sortedcontainers
SQLAlchemy
sqlparse
statsmodels
stumpy
tabulate
tblib
tenacity
tensorboard
# tensorboard-data-server
tensorflow
tensorflow-estimator
# tensorflow-macos
termcolor
# TestSlide
threadpoolctl
toml
# tomli
tomlkit
toolz
tornado
tqdm
tsfresh
typeguard
typing-inspect
typing_extensions
tzdata
urllib3
Werkzeug
wrapt
xgboost
zict
# zipp

